refactor: BoardCreator 인터페이스 생성 및 불필요한 Supplier 제거
